- Use the %ju conversion specifier for time_t and get rid of these non-portable (unsigned long) casts; replace some uint64_t with uintmax_t
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/branches/8.2.x@6781 82007160-df01-0410-b94d-b575c5fd34c7

+++ modules/core/m_bmask.c
@@ -136,8 +136,8 @@ ms_bmask(struct Client *source_p, int parc, char *parv[])
     sendto_channel_local(NULL, chptr, 0, 0, 0, "%s %s", modebuf, parabuf);
   }
 
-  sendto_server(source_p, 0, 0, ":%s BMASK %lu %s %s :%s",
-                source_p->id, (unsigned long)chptr->creationtime, chptr->name,
+  sendto_server(source_p, 0, 0, ":%s BMASK %ju %s %s :%s",
+                source_p->id, chptr->creationtime, chptr->name,
                 parv[3], parv[4]);
   return 0;
 }
